requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.68.2
firebase-admin==5.3.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
gridfs-fuse==0.1.2.dev1
grpcio==1.48.1
grpcio-status==1.48.1
h11==0.13.0
httplib2==0.20.4
idna==3.3
joblib==1.1.0
llfuse==1.4.2
msgpack==1.0.4
proto-plus==1.22.1
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pymongo==4.0.2
pyodbc==4.0.32
pyparsing==3.0.9
python-multipart==0.0.5
requests==2.27.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.31
starlette==0.14.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.15.0
